v1.0

+++ backend/app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Date, DateTime, Text
+from sqlalchemy import Column, Integer, String, Float, Date, DateTime, Text, BigInteger
 from sqlalchemy.sql import func
 from database import Base
 
@@ -6,7 +6,7 @@ class Expense(Base):
     __tablename__ = "expenses"
     
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, index=True)
+    user_id = Column(BigInteger, index=True)  # ← BIGINT
     amount = Column(Float, nullable=False)
     category = Column(String, index=True)
     description = Column(Text, nullable=True)
@@ -17,7 +17,7 @@ class Income(Base):
     __tablename__ = "income"
     
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, index=True)
+    user_id = Column(BigInteger, index=True)  # ← BIGINT
     amount = Column(Float, nullable=False)
     source = Column(String)
     description = Column(Text, nullable=True)
